+++ src/plugins/__init__.py
@@ -9,6 +9,17 @@ from .agent_tools import (
 )
 from .constants import PLUGIN_APPLICATION_VERSION
 from .errors import PluginContextClosedError, PluginError, PluginRegistryError
+from .event_hooks import (
+    EVENT_HOOK_NAMES,
+    EVENT_HOOK_SCHEMA_VERSION,
+    EventHook,
+    EventHookRegistration,
+    PluginEvent,
+    dispatch_analysis_event,
+    dispatch_market_review_event,
+    event_hook_extension_contract,
+    validate_event_hook_registration,
+)
 from .loader import ExternalPluginLoader, ExternalPluginResult
 from .manager import (
     PluginManager,
@@ -35,6 +46,10 @@ from .registry import (
 __all__ = [
     "AgentToolRegistrationBackend",
     "EXTENSION_POINTS",
+    "EVENT_HOOK_NAMES",
+    "EVENT_HOOK_SCHEMA_VERSION",
+    "EventHook",
+    "EventHookRegistration",
     "ExternalPluginLoader",
     "ExternalPluginResult",
     "ExtensionContract",
@@ -44,6 +59,7 @@ __all__ = [
     "JSONValue",
     "NativeRegistrationBackend",
     "Plugin",
+    "PluginEvent",
     "PLUGIN_APPLICATION_VERSION",
     "PluginContext",
     "PluginContextClosedError",
@@ -58,5 +74,9 @@ __all__ = [
     "RegistrationHandle",
     "build_agent_tool_extension_registry",
     "default_extension_contracts",
+    "dispatch_analysis_event",
+    "dispatch_market_review_event",
+    "event_hook_extension_contract",
     "validate_agent_tool_definition",
+    "validate_event_hook_registration",
 ]